% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_collection.R
\name{bdr_collection}
\alias{bdr_collection}
\title{Parse BDR Collection}
\usage{
bdr_collection(url)
}
\arguments{
\item{url}{A character URL pointing to a JSON output}
}
\value{
A \code{tibble} of information of collections
}
\description{
Parse BDR Collection
}
\examples{
library(dplyr)
df = bdr_collections()
df
url = df \%>\%
    filter(grepl("Bactrian ", name)) \%>\%
    pull(json_uri)
coll = bdr_collection(url)
testthat::expect_is(coll, "tbl")
item_urls =coll$items[[1]]$json_uri

}
