% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_collections.R
\name{bdr_collections}
\alias{bdr_collections}
\title{BDR Collections}
\usage{
bdr_collections()
}
\value{
A \code{tibble} of information of collections
}
\description{
BDR Collections
}
\examples{
library(dplyr)
df =bdr_collections()
df
uri = df \%>\%
    filter(grepl("Bactrian ", name)) \%>\%
    pull(json_uri)
}
