% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_item_link.R
\name{bdr_item_download_link}
\alias{bdr_item_download_link}
\title{Parse BDR Item Downlaod Link}
\usage{
bdr_item_download_link(url)
}
\value{
A \code{character} vector of links.
}
\description{
Parse BDR Item Downlaod Link
}
\examples{
library(dplyr)
df = bdr_collections()
df
coll_url = df \%>\%
    filter(grepl("Bactrian ", name)) \%>\%
    pull(json_uri)
coll = bdr_collection(coll_url)
testthat::expect_is(coll, "tbl")
url =coll$items[[1]]$json_uri[1]
link = bdr_item_download_link(url)
tfile = tempfile(fileext = ".zip")
dl = httr::GET(link,
httr::write_disk(tfile),
if (interactive()) httr::progress())
unz = unzip(tfile)

}
