% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdr_parse_url.R
\name{bdr_parse_url}
\alias{bdr_parse_url}
\title{BDR Parse JSON URL}
\usage{
bdr_parse_url(url)
}
\arguments{
\item{url}{A character URL pointing to a JSON output}
}
\value{
A list of the request and a
\code{list} of content.
}
\description{
BDR Parse JSON URL
}
\examples{
library(dplyr)
df =bdr_collections()
df
url = df \%>\%
    filter(grepl("Bactrian ", name)) \%>\%
    pull(json_uri)
res = bdr_parse_url(url)
res$data
}
