% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFigAntallProsent.R
\name{rFigAntallProsent}
\alias{rFigAntallProsent}
\title{Figur for antall og prosent}
\usage{
rFigAntallProsent(vs = TRUE, figdata = figdata, figdata2 = figdata2,
  yAksen11 = yAksen, N = N, andreN = NULL, sykehusNavn = sykehusNavn,
  sykehusAndre = NULL, ymax = ymax, yLab = yLab, xLab = xLab,
  figsubT = figsubT, txtSpace = txtSpace, col1 = col1, col2 = col2,
  coll = coll)
}
\arguments{
\item{vs}{Sammenligne lokal og landet}

\item{figdata}{Lokal sykehus datasett for figur}

\item{figdata2}{Øvrige sykehus datasettet for figur}

\item{N}{Antall for lokal sykehus inkludert i analysen}

\item{andreN}{Antall for øvrige sykehus inkludert i analysen}

\item{sykehusNavn}{Tekst til legend for lokal sykehus}

\item{sykehusAndre}{Tekst til legend for øvrige sykehus}

\item{ymax}{Max for y-aksen}

\item{yLab}{Label for y-aksen}

\item{xLab}{Label for x-aksen}

\item{figsubT}{Tekst for title til figuren}

\item{txtSpace}{Linje å lage til figur title}

\item{col1}{Farge for søyle}

\item{col2}{Farge for point}

\item{coll}{Farge for linje i figur}

\item{Variabel}{Variabel for x-aksen}

\item{yAksen}{Variabel for y-aksen}
}
\description{
Figur for antall og prosent
}

