% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfigur.R
\name{rfigur}
\alias{rfigur}
\title{Beregn og Figur}
\usage{
rfigur(data = NULL, sykehus = NULL, rapValg = NULL, yAksen = 2,
  xScale = 2, figT = "none", figTxt = "none", xBreaks = NULL,
  save = NULL, format = "pdf", path = NULL)
}
\arguments{
\item{data}{Datasettet}

\item{sykehus}{Utvalgte sykehus (lokal)}

\item{rapValg}{Rapport valgt dvs. 1:Lokal, 2:Landet og 3:Lokal vs. øvrige}

\item{yAksen}{Y aksen for figuren}

\item{xScale}{Skale å velge dvs. 1 = kontinuell og 2 = kategorisk}

\item{figT}{Title for figur}

\item{figTxt}{Tekst info om inholde i data for figuren}

\item{xBreaks}{Navnene for variabler i x-aksen}

\item{save}{Filnavn for å lage figuren}

\item{format}{Figur format i.e "pdf", "jpg" eller "png"}

\item{path}{Hvor skal figuren lagres evt. hjemme område}
}
\description{
Her skal det beregne antall, andel eller andre typer beregninger som trenges før figuren skal lages
}

