% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rselect.R
\name{rselect}
\alias{rselect}
\title{Velger variabel til figur}
\usage{
rselect(data = NULL, valgtVar = NULL, figTx = "none", minx = 0,
  maxx = 99)
}
\arguments{
\item{data}{Datasettet}

\item{valgtVar}{Utvalgte variabel til figuren}

\item{figTx}{Tekst som beskriver data i figuren om funksjonen brukes direkte}

\item{minx}{Minimum tall for x-aksen}

\item{maxx}{Maksimum tall for x-aksen}
}
\description{
Her skal utvalgte variabel videre behandlet for figuren er lages
}

