% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclean.R
\name{rclean}
\alias{rclean}
\title{Clean dataset}
\usage{
rclean(data = NULL)
}
\arguments{
\item{data}{Datasettet som er ferdig konvertert til R format}
}
\description{
Denne funksjonen skal rense datasettet og lage nye variabler hvis nødvendig
}

