% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfile.R
\name{rfile}
\alias{rfile}
\title{Konvertere datasettet til R format}
\usage{
rfile(filnavn = NULL)
}
\arguments{
\item{filnavn}{Filnavn hentet fra register database}
}
\value{
Et dataset i R data.frame format
}
\description{
Denne funksjon skal konvertere datafil fra forskjellige formater dvs. sav, dat, txt og csv
til en format som R kan håndtere. Excel fil er ikke akseptert.
}

